Change generated files to be C instead of headers.
Remove AdobeCA.p7c from autogenerated files. Just include the
array in the source.

Simplifies makefile dependencies and makes the sizes of each
bit of data easier to look at.

It also paves the way for eventually using objcopy to create binary
objects for the fonts instead of needing to use hexdump.